backend finished

+++ user/views.py
@@ -4,7 +4,6 @@ from rest_framework.decorators import api_view
 from rest_framework.response import Response
 from .serializers import UserSerializer
 from rest_framework.authtoken.models import Token
-from django.utils import timezone
 from django.contrib.auth.models import User
 from django.db import IntegrityError
 # Create your views here.
@@ -32,6 +31,11 @@ def signup(request):
         user = User.objects.create_user(email,email,password)
         user.first_name = name
         user.save()
+        token, created = Token.objects.get_or_create(user=user)
+        user_serializer = UserSerializer(user)
+        response = user_serializer.data
+        response['token']=token.key
+        return Response(response,status=200)
         return Response({"status":"success"},status=201)
     except IntegrityError as e:
         return Response({"status":"fail","message":"Username is not available"},status=400)
@@ -42,8 +46,9 @@ def signup(request):
 def logout(request):
     if request.user:
         try:
-            token = Token.objects.filter(user=request.user).first()
-            token.delete()
+            token = Token.objects.filter(user=request.user.id).first()
+            if token:
+                token.delete()
         except Exception as e:
             return Response({"status":"fail","message":str(e)},status=500)
     return Response({"status":"success"},status=200)
